+++ src/oci/core/models/subnet.py
@@ -9,7 +9,7 @@ from oci.decorators import init_model_state_from_kwargs
 @init_model_state_from_kwargs
 class Subnet(object):
     """
-    A logical subdivision of a VCN. Each subnet exists in a single Availability Domain and
+    A logical subdivision of a VCN. Each subnet exists in a single availability domain and
     consists of a contiguous range of IP addresses that do not overlap with
     other subnets in the VCN. Example: 172.16.1.0/24. For more information, see
     `Overview of the Networking Service`__ and
@@ -18,6 +18,9 @@ class Subnet(object):
     To use any of the API operations, you must be authorized in an IAM policy. If you're not authorized,
     talk to an administrator. If you're an administrator who needs to write policies to give users access, see
     `Getting Started with Policies`__.
+
+    **Warning:** Oracle recommends that you avoid using any confidential information when you
+    supply string values using the API.
 
     __ https://docs.us-phoenix-1.oraclecloud.com/Content/Network/Concepts/overview.htm
     __ https://docs.us-phoenix-1.oraclecloud.com/Content/Network/Tasks/managingVCNs.htm
@@ -185,7 +188,7 @@ class Subnet(object):
     def availability_domain(self):
         """
         **[Required]** Gets the availability_domain of this Subnet.
-        The subnet's Availability Domain.
+        The subnet's availability domain.
 
         Example: `Uocm:PHX-AD-1`
 
@@ -199,7 +202,7 @@ class Subnet(object):
     def availability_domain(self, availability_domain):
         """
         Sets the availability_domain of this Subnet.
-        The subnet's Availability Domain.
+        The subnet's availability domain.
 
         Example: `Uocm:PHX-AD-1`