+++ src/oci/core/models/vcn.py
@@ -9,12 +9,15 @@ from oci.decorators import init_model_state_from_kwargs
 @init_model_state_from_kwargs
 class Vcn(object):
     """
-    A Virtual Cloud Network (VCN). For more information, see
+    A virtual cloud network (VCN). For more information, see
     `Overview of the Networking Service`__.
 
     To use any of the API operations, you must be authorized in an IAM policy. If you're not authorized,
     talk to an administrator. If you're an administrator who needs to write policies to give users access, see
     `Getting Started with Policies`__.
+
+    **Warning:** Oracle recommends that you avoid using any confidential information when you
+    supply string values using the API.
 
     __ https://docs.us-phoenix-1.oraclecloud.com/Content/Network/Concepts/overview.htm
     __ https://docs.us-phoenix-1.oraclecloud.com/Content/Identity/Concepts/policygetstarted.htm