+++ src/oci/core/models/volume_group.py
@@ -9,7 +9,11 @@ from oci.decorators import init_model_state_from_kwargs
 @init_model_state_from_kwargs
 class VolumeGroup(object):
     """
-    Specifies a volume group which is a collection of volumes. For more information, see `Volume Groups`__.
+    Specifies a volume group which is a collection of
+    volumes. For more information, see `Volume Groups`__.
+
+    **Warning:** Oracle recommends that you avoid using any confidential information when you
+    supply string values using the API.
 
     __ https://docs.us-phoenix-1.oraclecloud.com/Content/Block/Concepts/volumegroups.htm
     """