+++ src/oci/database/models/create_data_guard_association_to_existing_db_system_details.py
@@ -68,7 +68,7 @@ class CreateDataGuardAssociationToExistingDbSystemDetails(CreateDataGuardAssocia
     def peer_db_system_id(self):
         """
         Gets the peer_db_system_id of this CreateDataGuardAssociationToExistingDbSystemDetails.
-        The `OCID`__ of the DB System to create the standby database on.
+        The `OCID`__ of the DB system to create the standby database on.
 
         __ https://docs.us-phoenix-1.oraclecloud.com/Content/General/Concepts/identifiers.htm
 
@@ -82,7 +82,7 @@ class CreateDataGuardAssociationToExistingDbSystemDetails(CreateDataGuardAssocia
     def peer_db_system_id(self, peer_db_system_id):
         """
         Sets the peer_db_system_id of this CreateDataGuardAssociationToExistingDbSystemDetails.
-        The `OCID`__ of the DB System to create the standby database on.
+        The `OCID`__ of the DB system to create the standby database on.
 
         __ https://docs.us-phoenix-1.oraclecloud.com/Content/General/Concepts/identifiers.htm